RF | move methods from class DisableEnableProjectRDVTest in the class _FreestyleTest (RDV)
